+++ R/guide_axis_truncated.R
@@ -1,9 +1,12 @@
 # User function -----------------------------------------------------------
 
-#' Axis guide with truncated line.
+#' Axis guide with truncated line
 #'
 #' This axis guide is similar to the normal axis guides for position scales, but
-#' can shorten the axis line that is being drawn.
+#' can shorten the axis line that is being drawn. The `guide_axis_colour()`
+#' function is the same but with different defaults for the truncation that do
+#' not truncate the axis. Axis truncation and recolouring is supported
+#' throughout axes in ggh4x.
 #'
 #' @inheritParams ggplot2::guide_axis
 #' @param trunc_lower,trunc_upper The lower and upper range of the truncated
@@ -14,6 +17,9 @@
 #'   mapped positions as `numeric`.
 #' * A `numeric` value in data units for the lower and upper boundaries.
 #' * A `unit` object.
+#' @param colour,color A `character(1)` with a valid colour for colouring the
+#'   axis text, axis ticks and axis line. Overrules the colour assigned by the
+#'   theme.
 #'
 #' @return An *axis_ggh4x* guide class object.
 #' @export
@@ -45,16 +51,23 @@
 #'   trunc_lower = function(x) {x - 0.2},
 #'   trunc_upper = function(x) {x + 0.2}
 #' ))
+#'
+#' # Recolouring the axes outside the theme
+#' p + guides(x = guide_axis_colour(colour = "red"),
+#'            y = guide_axis_colour(colour = "blue"))
 guide_axis_truncated <- function(
   title = waiver(),
   check.overlap = FALSE,
   angle = NULL,
   n.dodge = 1,
   order = 0,
+  colour = NULL,
+  color = NULL,
   trunc_lower = min,
   trunc_upper = max,
   position = waiver()
 ) {
+  colour <- color %||% colour
   check_trunc_arg(trunc_lower, trunc_upper)
   structure(
     list(
@@ -65,6 +78,7 @@ guide_axis_truncated <- function(
       order = order,
       trunc_lower = trunc_lower,
       trunc_upper = trunc_upper,
+      colour = colour,
       position = position,
       available_aes = c("x", "y"),
       name = "axis"
@@ -72,6 +86,44 @@ guide_axis_truncated <- function(
     class = c("guide", "axis_ggh4x", "axis")
   )
 }
+
+#' @rdname guide_axis_truncated
+#' @export
+guide_axis_colour <- function(
+  title = waiver(),
+  check.overlap = FALSE,
+  angle = NULL,
+  n.dodge = 1,
+  order = 0,
+  colour = NULL,
+  color = NULL,
+  trunc_lower = NULL,
+  trunc_upper = NULL,
+  position = waiver()
+) {
+  colour <- color %||% colour
+  check_trunc_arg(trunc_lower, trunc_upper)
+  structure(
+    list(
+      title = title,
+      check.overlap = check.overlap,
+      angle = angle,
+      n.dodge = n.dodge,
+      order = order,
+      trunc_lower = trunc_lower,
+      trunc_upper = trunc_upper,
+      colour = colour,
+      position = position,
+      available_aes = c("x", "y"),
+      name = "axis"
+    ),
+    class = c("guide", "axis_ggh4x", "axis")
+  )
+}
+
+#' @rdname guide_axis_truncated
+#' @export
+guide_axis_color <- guide_axis_colour
 
 # Internals ---------------------------------------------------------------
 
@@ -104,7 +156,8 @@ guide_gengrob.axis_ggh4x <- function(guide, theme) {
     check.overlap = guide$check.overlap,
     angle = guide$angle,
     n.dodge = guide$n.dodge,
-    trunc = guide$trunc
+    trunc = guide$trunc,
+    colour = guide$colour
   )
 }
 
@@ -118,13 +171,14 @@ draw_axis_ggh4x <- function(
   check.overlap,
   angle = NULL,
   n.dodge = 1,
-  trunc
+  trunc,
+  colour = NULL
 ) {
   axis_position <- match.arg(substr(axis_position, 1, 1),
                              c("t", "b", "r", "l"))
   aes <- if (axis_position %in% c("t", "b")) "x" else "y"
 
-  elements <- build_axis_elements(axis_position, angle, theme)
+  elements <- build_axis_elements(axis_position, angle, theme, colour)
   params <- setup_axis_params(axis_position)
   line_grob <- build_trunc_axis_line(elements$line, params, trunc)
 
@@ -239,3 +293,4 @@ check_trunc_arg <- function(lower, upper) {
     }
   }
 }
+